requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
onnxruntime==1.16.3
onnxruntime-gpu==1.16.3
torch==2.1.1
torchvision==0.16.1
boto3==1.34.0
evidently==0.4.9
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
pandas==2.1.4
pydantic==2.5.1
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0
aiohttp==3.9.1 